inttests regresssion tests store test desc in resp filename
Refactor inttests/main.go and main_test.go so that the test
description is part of the filename (in addition to the files
sha1 hash) for the response files used to check regression.

New test cases for which there is no repsonse file will
always fail, and will prompt you to re-run main.go to record
the response file.

This change needed for easier maintenance.