get-readme-scaffold.sh makes readme scaffold from go doc output
shell script to crawl project looking for public go code. which is
then documented with `go doc` and appended to readme-scaffold.md
as preformatted text. Any README.md files found in package
directories will be linked to in readme-scaffold.md.

This change needed to make it easier to document this project.