add censustable param to geodata
Add new parameter 'censustable' (string) to
geodata.App. This represents a short nomis code /
nomis table, e.g. a grouping of census data categories,
e.g. QS101EW. When used in a query, it will be expanded
to all cateogories that are part of that table.
This change needed to simplify API querying for all
categories within a table.

Tests added in geodata_test.go.