default DSN, SetupDBOnceOnly db setup function, both for comptests
- Add DefaultDSN constant to comptests/docker.go. This can be
imported into component tests, and avoids tests thinking they
can set different dsn's for the test db in each test run (actually
only the first one to be used is valid)
- Add SetupDBOnceOnly function to model/provision.go. This is a
wrapper for SetupDB that is safe to call multiple times (checks
for existence of db and only calls SetupDB if not present). This
function can be called multiple times by different comptests
in the same test session.

These changes needed to allow multiple component tests to
run in the same test session.