blank geo response if no geometry available (e.g. E&W)
Return no data from geo endpoint if there is no geometry available
in the geo table for the requested geography (but there is a name
and a code). This covers cases like England and Wales, which
are geographies we hold data for but have no associated geometry.

This change needed to avoid error response when trying to marshal
blank geometry values to geoJSON.